무기 업그레이드 테이블 및 레벨업 리워드

+++ Assets/01.Scripts/Character/Player/Weapon/Module/AoEModule.cs
@@ -1,5 +1,5 @@
 ﻿[System.Serializable]
 public class AoEModule
 {
-    public float radius;
+    public int radius;
 }
